split up instances module
